UScensuscounty20 This package contains county level data for the US census 2020 redistricting files and shape files.

FILEID

File Identification

STUSAB

State/US-Abbreviation (USPS)

SUMLEV

Summary Level=050 for county

GEOVAR

Geographic Variant

GEOCOMP

Geographic Component

CHARITER

Characteristic Iteration

CIFSN

Characteristic Iteration File Sequence Number

LOGRECNO

Logical Record Number

GEOID

Geographic Record Identifier

GEOCODE

Geographic Code Identifier

REGION

Region

DIVISION

Division

STATE

State

COUNTY

County

COUNTYCC

FIPS County class code

COUNTYNS

County (NS)

COUSUB

County Sub Divison (FIPS)

COUSUBCC

FIPS County Subdivision Class Code

COUSUBNS

County Subdivision (NS)

SUBMCD

Subminor Civil Division (FIPS)

SUBMCDCC

FIPS Subminor Civil Division Class Code

SUBMCDNS

Subminor Civil Division (NS)

ESTATE

Estate (FIPS)

ESTATECC

FIPS Estate Class Code

ESTATENS

Estate (NS)

CONCIT

Consolidated City (FIPS)

CONCITCC

FIPS Consolidated City Class Code

CONCITNS

Consolidated City (NS)

PLACE

Place (FIPS)

PLACECC

FIPS Place Class Code

PLACENS

Place (NS)

TRACT

Census Tract

BLKGRP

Block Group

BLOCK

Block

AIANHH

American Indian Area/Alaska Native Area/Hawaiian Home Land (Census)

AIHHTLI

American Indian Trust Land/Hawaiian Home Land Indicator

AIANHHFP

American Indian Area/Alaska Native Area/Hawaiian Home Land (FIPS)

AIANHHCC

FIPS American Indian Area/Alaska Native Area/Hawaiian Home Land Class Code

AIANHHNS

American Indian Area/Alaska Native Area/Hawaiian Home Land (NS)

AITS

American Indian Tribal Subdivision (Census)

AITSFP

American Indian Tribal Subdivision (FIPS)

AITSCC

FIPS American Indian Tribal Subdivision Class Code

AITSNS

American Indian Tribal Subdivision (NS)

TTRACT

Tribal Census Tract

TBLKGRP

Tribal Block Group

ANRC

Alaska Native Regional Corporation (FIPS)

ANRCCC

FIPS Alaska Native Regional Corporation Class Code

ANRCNS

Alaska Native Regional Corporation (NS)

CBSA

Metropolitan Statistical Area/Micropolitan Statistical Area

MEMI

Metropolitan/Micropolitan Indicator

CSA

Combine Statistical Area

METDIV

Metropolitan Division

NECTA

New England City and Town Area

NMEMI

NECTA Metropolitan/Micropolitan Indicator

CNECTA

Combined New England City and Town Area

NECTADIV

New England city and town division

CBSAPCI

Metropolitan Statistical Area/Micropolitan Statistical Area Principal City Indicator

NECTAPCI

New England City and Town Area Principal City Indicator

UA

Urban Area

UATYPE

Urban Area Type

UR

Urban/Rural

CD116

Congressional District (116th)

CD118

Congressional District (118th)

CD119

Congressional District (119th)

CD120

Congressional District (120th)

CD121

Congressional District (121th)

SLDU18

State Legislative District (Upper Chamber) (2018)

SLDU22

State Legislative District (Upper Chamber) (2022)

SLDU24

State Legislative District (Upper Chamber) (2024)

SLDU26

State Legislative District (Upper Chamber) (2026)

SLDU28

State Legislative District (Upper Chamber) (2028)

SLDL18

State Legislative District (Lower Chamber) (2018)

SLDL22

State Legislative District (Lower Chamber) (2022)

SLDL24

State Legislative District (Lower Chamber) (2024)

SLDL26

State Legislative District (Lower Chamber) (2026)

SLDL28

State Legislative District (Lower Chamber) (2028)

VTD

Voting District

VTDI

Voting District Indicator

ZCTA

ZIP Code Tabulation Area (5-Digit)

SDELM

School District (Elementary)

SDSEC

School District (Secondary)

SDUNI

School District (Unifield)

PUMA

Public Use Microdata Area

AREALAND

Area(Land)

AREAWATR

Area(Water)

BASENAME

Area Base Name

NAME

Area Name-Legal/Statistical Area Description (LSAD) Term-Part Indicator

FUNCSTAT

Functional Status Code

GCUNI

Geographic Change User Note Indicator

POP100

Population Count (100%)

HU100

Housing Unit Count (100%)

INTPTLAT

Internal Point (Latitude)

INTPTLON

Internal Point (Longitude)

LSADC

Legal/Statistical Area Description Code

PARTFLAG

Part Flag

UGA

Urban Growth Area

P0010001

Total Population

P0010002

Population of single race

P0010003

Population of white race alone

P0010004

Population of Black or African american population alone

P0010005

American Indian and Alaskan native alone

P0010006

Asian Alone

P0010007

Native Hawaiian and Other Pacific Islander alone

P0010008

Some other race alone

P0010009

Population of two or more races

P0010010

Population of two races

P0010011

White and black or african american population

P0010012

White and American Indian or alaskan native population

P0010013

White and Asian population

P0010014

White and Native Hawaiian and other Pacific Islander population

P0010015

White and some other race

P0010016

Black or African American and American Indian and Alaska Native Population

P0010017

Black or African American and Asian Population

P0010018

Black or African American and Native Hawaiian and other Pacific Islander Population

P0010019

Black or African American and some other race population

P0010020

American Indian and Alaska Native and Asian Population

P0010021

American Indian and Alaska Native and Native Hawaiian and other pacific islander population

P0010022

American Indian and Alaska Native and some other race population

P0010023

Asian; Native Hawaiian and Other Pacific Islander

P0010024

Asian and some other race population

P0010025

Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010026

Population of three races

P0010027

white; Black or African American; American Indian and Alaska Native population

P0010028

White; Black or african american; Asian Population

P0010029

white; Black or African American; Native Hawaiian and other pacific islander population

P0010030

White; Black or African American; Some Other Race population

P0010031

White; American Indian and Alaska Native; Asian population

P0010032

White; American Indian and Alaska Native; Native Hawaiian and other pacific islander population

P0010033

White; American Indian and Alaska Native; Some other race population

P0010034

White; Asian; Native Hawaiian and other pacific Islander Population

P0010035

White; Asian; Some other race population

P0010036

White; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010037

Black or African American; American Indian and Alaska Native; Asian Population

P0010038

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander population

P0010039

Black or African American; American Indian and Alaska Native; Some Other Race population

P0010040

Black or African American; Asian; Native Hawaiian and Other Pacific Islander Population

P0010041

Black or African American; Asian; Some Other Race population

P0010042

Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010043

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander Population

P0010044

American Indian and Alaska Native; Asian; Some Other Race population

P0010045

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010046

Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010046

Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010047

Population of four races

P0010048

White; Black or African American; American Indian and Alaska Native; Asian

P0010049

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0010050

White; Black or African American; American Indian and Alaska Native; Some Other Race

P0010051

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0010052

White; Black or African American; Asian; Some Other Race

P0010053

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010054

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0010055

White; American Indian and Alaska Native; Asian; Some Other Race

P0010056

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010057

White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010058

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0010059

Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0010060

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010061

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010062

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010063

Population of five races

P0010064

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0010065

White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0010066

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010067

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010068

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010069

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010070

Population of six races

P0010071

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020001

Total:

P0020002

Hispanic or Latino

P0020003

Not Hispanic or Latino

P0020004

Population of one race

P0020005

White alone

P0020006

Black or african american alone

P0020007

American Indian or Alaskan Native alone

P0020008

Asian Alone

P0020009

Native Hawaiian and Other Pacific Islander alone

P0020010

Some other race alone

P0020011

Population of two or more races

P0020012

Population of two races

P0020013

White; Black or African American

P0020014

White; American Indian and Alaska Native

P0020015

White; Asian

P0020016

White; Native Hawaiian and Other Pacific Islander

P0020017

White; Some other race

P0020018

Black or African American; American Indian and Alaska Native

P0020019

Black or African American; Asian

P0020020

Black or African American; Native Hawaiian and Other Pacific Islander

P0020021

Black or African American; Some Other Race

P0020022

American Indian and Alaska Native; Asian

P0020023

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020024

American Indian and Alaska Native; Some Other Race

P0020025

Asian; Native Hawaiian and Other Pacific Islander

P0020026

Asian; Some Other Race

P0020027

Native Hawaiian and Other Pacific Islander; Some Other Race

P0020028

Population of three races

P0020029

White; Black or African American; American Indian and Alaska Native

P0020030

White; Black or African American; Asian

P0020031

White; Black or African American; Native Hawaiian and Other Pacific Islander

P0020032

White; Black or African American; Some Other Race

P0020033

White; American Indian and Alaska Native; Asian

P0020034

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020035

White; American Indian and Alaska Native; Some Other Race

P0020036

White; Asian; Native Hawaiian and Other Pacific Islander

P0020037

White; Asian; Some Other Race

P0020038

White; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020039

Black or African American; American Indian and Alaska Native; Asian

P0020040

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020021 Black or African American; Some Other Race
P0020041

Black or African American; American Indian and Alaska Native; Some Other Race

P0020042

Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0020043

Black or African American; Asian; Some Other Race

P0020044

Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020045

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020046

American Indian and Alaska Native; Asian; Some Other Race

P0020047

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020048

Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020049

Population of four races;

P0020050

White; Black or African American; American Indian and Alaska Native; Asian

P0020051

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020052

White; Black or African American; American Indian and Alaska Native; Some Other Race

P0020053

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0020054

White; Black or African American; Asian; Some Other Race

P0020055

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020056

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020057

White; American Indian and Alaska Native; Asian; Some Other Race

P0020058

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020059

White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020060

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020041 Black or African American; American Indian and Alaska Native; Some Other Race
P0020061

Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0020062

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020063

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020064

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020065

Population of five races

P0020066

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020067

White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0020068

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020069

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020070

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020071

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020072

Population of six races

P0020073

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0030001

Total

P0030002

Population of one race

P0030003

White alone

P0030004

Black or african american alone

P0030005

American Indian and Alaska native alone

P0030006

Asian Alone

P0030007

Native Hawaiian and pacific islander alone

P0030008

Some other race alone

P0030009

Population of two or more races

P0030010

Population of two races

P0030011

White; Black or African American

P0030012

White; American Indian and Alaska Native

P0030013

White; Asian

P0030014

White; Native Hawaiian and Other Pacific Islander

P0030015

White; Some other race

P0030016

Black or African American; American Indian and Alaska Native

P0030017

Black or African American; Asian

P0030018

Black or African American; Native Hawaiian and Other Pacific Islander

P0030019

Black or African American; Some other race

P0030020

American Indian and Alaska Native; Asian

P0030021

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030022

American Indian and Alaska Native; Some other race

P0030023

Asian; Native Hawaiian and Other Pacific Islander

P0030024

Asian; Some other race

P0030025

Native Hawaiian and Other Pacific Islander; Some other race

P0030026

Population of three races

P0030027

White; Black or African American; American Indian and Alaska Native

P0030028

White; Black or African American; Asian

P0030029

White; Black or African American; Native Hawaiian and Other Pacific Islander

P0030030

White; Black or African American; Some other race

P0030031

White; American Indian and Alaska Native; Asian

P0030032

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030033

White; American Indian and Alaska Native; Some other race

P0030034

White; Asian; Native Hawaiian and Other Pacific Islander

P0030035

White; Asian; Some other race

P0030036

White; Native Hawaiian and Other Pacific Islander; Some other race

P0030037

Black or African American; American Indian and Alaska Native; Asian

P0030038

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030039

Black or African American; American Indian and Alaska Native; Some other race

P0030040

Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0030041

Black or African American; Asian; Some other race

P0030042

Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0030043

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030044

American Indian and Alaska Native; Asian; Some other race

P0030045

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030046

Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030047

Population of four races

P0030048

White; Black or African American; American Indian and Alaska Native; Asian

P0030049

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030050

White; Black or African American; American Indian and Alaska Native; Some other race

P0030051

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0030052

White; Black or African American; Asian; Some other race

P0030053

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0030054

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030055

White; American Indian and Alaska Native; Asian; Some other race

P0030056

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030057

White; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030058

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030059

Black or African American; American Indian and Alaska Native; Asian; Some other race

P0030060

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030061

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030062

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030063

Population of five races

P0030064

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030065

White; Black or African American; American Indian and Alaska Native; Asian; Some other race

P0030066

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030067

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030068

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030069

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030070

Population of six races

P0030071

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040001

Total

P0040002

Hispanic or latino

P0040003

Not hispanic or latino

P0040004

Population of one race

P0040005

White alone

P0040006

Black or african american alone

P0040007

American Indian and alaska native alone

P0040008

Asian Alone

P0040009

Native Hawaiian and pacific islander alone

P0040010

Some other race alone

P0040011

Population of two or more races

P0040012

Population of two races

P0040013

White; Black or African American

P0040014

White; American Indian and Alaska Native

P0040015

White; Asian

P0040016

White; Native Hawaiian and Other Pacific Islander

P0040017

White; Some other race

P0040018

Black or African American; American Indian and Alaska Native

P0040019

Black or African American; Asian

P0040020

Black or African American; Native Hawaiian and Other Pacific Islander

P0040021

Black or African American; Some other race

P0040022

American Indian and Alaska Native; Asian

P0040023

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040024

American Indian and Alaska Native; Some other race

P0040025

Asian; Native Hawaiian and Other Pacific Islander

P0040026

Asian; SOme other race

P0040027

Native Hawaiian and Other Pacific Islander; Some other race

P0040028

Population of three races

P0040029

White; Black or African American; American Indian and Alaska Native

P0040030

White; Black or African American; Asian

P0040031

White; Black or African American; Native Hawaiian and Other Pacific Islander

P0040032

White; Black or African American; Some other race

P0040033

White; American Indian and Alaska Native; Asian

P0040034

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040035

White; American Indian and Alaska Native; Some other race

P0040036

White; Asian; Native Hawaiian and Other Pacific Islander

P0040037

White; Asian; Some other race

P0040038

White; Native Hawaiian and Other Pacific Islander; Some other race

P0040039

Black or African American; American Indian and Alaska Native; Asian

P0040040

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040041

Black or African American; American Indian and Alaska Native; Some other race

P0040042

Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0040043

Black or african american; Asian; some other race

P0040044

Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0040045

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040046

American Indian and Alaska Native; Asian; Some other race

P0040047

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040048

Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040049

Population of four races

P0040050

White; Black or African American; American Indian and Alaska Native; Asian

P0040051

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040052

White; Black or African American; American Indian and Alaska Native; Some other race

P0040053

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0040054

White; Black or African American; Asian; Some other race

P0040055

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0040056

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040057

White; American Indian and Alaska Native; Asian; Some other race

P0040058

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040059

White; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040060

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040061

Black or African American; American Indian and Alaska Native; Asian; Some other race

P0040062

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040063

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040064

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040065

Population of five races

P0040066

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040067

White; Black or African American; American Indian and Alaska Native; Asian; Some other race

P0040068

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040069

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040070

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040071

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040072

Popukation of six races

P0040073

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

H0010001

Total

H0010002

Occupied

H0010003

Vacant

P0050001

Total

P0050002

Instituonalized population

P0050003

Correctional facilities for adults

P0050004

Juvenile facilities

P0050005

Nursing facilities/Skilled-nursing facilities

P0050006

Other institutional facilities

P0050007

Noninstitutionalized population:

P0050008

College/University student housing

P0050009

Military quarters

P0050010

Other noninstitutional facilities

MTFCC20

MAT/TIGER Feature Class Code

geometry

Geometric data for the shape. (Point, Multipoint, Polyline, Polygon,etc)

UScensuscounty20 This package contains county level data for the US census 2020 redistricting files and shape files.

FILEID

File Identification

STUSAB

State/US-Abbreviation (USPS)

SUMLEV

Summary Level=050 for county

GEOVAR

Geographic Variant

GEOCOMP

Geographic Component

CHARITER

Characteristic Iteration

CIFSN

Characteristic Iteration File Sequence Number

LOGRECNO

Logical Record Number

GEOID

Geographic Record Identifier

GEOCODE

Geographic Code Identifier

REGION

Region

DIVISION

Division

STATE

State

COUNTY

County

COUNTYCC

FIPS County class code

COUNTYNS

County (NS)

COUSUB

County Sub Divison (FIPS)

COUSUBCC

FIPS County Subdivision Class Code

COUSUBNS

County Subdivision (NS)

SUBMCD

Subminor Civil Division (FIPS)

SUBMCDCC

FIPS Subminor Civil Division Class Code

SUBMCDNS

Subminor Civil Division (NS)

ESTATE

Estate (FIPS)

ESTATECC

FIPS Estate Class Code

ESTATENS

Estate (NS)

CONCIT

Consolidated City (FIPS)

CONCITCC

FIPS Consolidated City Class Code

CONCITNS

Consolidated City (NS)

PLACE

Place (FIPS)

PLACECC

FIPS Place Class Code

PLACENS

Place (NS)

TRACT

Census Tract

BLKGRP

Block Group

BLOCK

Block

AIANHH

American Indian Area/Alaska Native Area/Hawaiian Home Land (Census)

AIHHTLI

American Indian Trust Land/Hawaiian Home Land Indicator

AIANHHFP

American Indian Area/Alaska Native Area/Hawaiian Home Land (FIPS)

AIANHHCC

FIPS American Indian Area/Alaska Native Area/Hawaiian Home Land Class Code

AIANHHNS

American Indian Area/Alaska Native Area/Hawaiian Home Land (NS)

AITS

American Indian Tribal Subdivision (Census)

AITSFP

American Indian Tribal Subdivision (FIPS)

AITSCC

FIPS American Indian Tribal Subdivision Class Code

AITSNS

American Indian Tribal Subdivision (NS)

TTRACT

Tribal Census Tract

TBLKGRP

Tribal Block Group

ANRC

Alaska Native Regional Corporation (FIPS)

ANRCCC

FIPS Alaska Native Regional Corporation Class Code

ANRCNS

Alaska Native Regional Corporation (NS)

CBSA

Metropolitan Statistical Area/Micropolitan Statistical Area

MEMI

Metropolitan/Micropolitan Indicator

CSA

Combine Statistical Area

METDIV

Metropolitan Division

NECTA

New England City and Town Area

NMEMI

NECTA Metropolitan/Micropolitan Indicator

CNECTA

Combined New England City and Town Area

NECTADIV

New England city and town division

CBSAPCI

Metropolitan Statistical Area/Micropolitan Statistical Area Principal City Indicator

NECTAPCI

New England City and Town Area Principal City Indicator

UA

Urban Area

UATYPE

Urban Area Type

UR

Urban/Rural

CD116

Congressional District (116th)

CD118

Congressional District (118th)

CD119

Congressional District (119th)

CD120

Congressional District (120th)

CD121

Congressional District (121th)

SLDU18

State Legislative District (Upper Chamber) (2018)

SLDU22

State Legislative District (Upper Chamber) (2022)

SLDU24

State Legislative District (Upper Chamber) (2024)

SLDU26

State Legislative District (Upper Chamber) (2026)

SLDU28

State Legislative District (Upper Chamber) (2028)

SLDL18

State Legislative District (Lower Chamber) (2018)

SLDL22

State Legislative District (Lower Chamber) (2022)

SLDL24

State Legislative District (Lower Chamber) (2024)

SLDL26

State Legislative District (Lower Chamber) (2026)

SLDL28

State Legislative District (Lower Chamber) (2028)

VTD

Voting District

VTDI

Voting District Indicator

ZCTA

ZIP Code Tabulation Area (5-Digit)

SDELM

School District (Elementary)

SDSEC

School District (Secondary)

SDUNI

School District (Unifield)

PUMA

Public Use Microdata Area

AREALAND

Area(Land)

AREAWATR

Area(Water)

BASENAME

Area Base Name

NAME

Area Name-Legal/Statistical Area Description (LSAD) Term-Part Indicator

FUNCSTAT

Functional Status Code

GCUNI

Geographic Change User Note Indicator

POP100

Population Count (100%)

HU100

Housing Unit Count (100%)

INTPTLAT

Internal Point (Latitude)

INTPTLON

Internal Point (Longitude)

LSADC

Legal/Statistical Area Description Code

PARTFLAG

Part Flag

UGA

Urban Growth Area

P0010001

Total Population

P0010002

Population of single race

P0010003

Population of white race alone

P0010004

Population of Black or African american population alone

P0010005

American Indian and Alaskan native alone

P0010006

Asian Alone

P0010007

Native Hawaiian and Other Pacific Islander alone

P0010008

Some other race alone

P0010009

Population of two or more races

P0010010

Population of two races

P0010011

White and black or african american population

P0010012

White and American Indian or alaskan native population

P0010013

White and Asian population

P0010014

White and Native Hawaiian and other Pacific Islander population

P0010015

White and some other race

P0010016

Black or African American and American Indian and Alaska Native Population

P0010017

Black or African American and Asian Population

P0010018

Black or African American and Native Hawaiian and other Pacific Islander Population

P0010019

Black or African American and some other race population

P0010020

American Indian and Alaska Native and Asian Population

P0010021

American Indian and Alaska Native and Native Hawaiian and other pacific islander population

P0010022

American Indian and Alaska Native and some other race population

P0010023

Asian; Native Hawaiian and Other Pacific Islander

P0010024

Asian and some other race population

P0010025

Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010026

Population of three races

P0010027

white; Black or African American; American Indian and Alaska Native population

P0010028

White; Black or african american; Asian Population

P0010029

white; Black or African American; Native Hawaiian and other pacific islander population

P0010030

White; Black or African American; Some Other Race population

P0010031

White; American Indian and Alaska Native; Asian population

P0010032

White; American Indian and Alaska Native; Native Hawaiian and other pacific islander population

P0010033

White; American Indian and Alaska Native; Some other race population

P0010034

White; Asian; Native Hawaiian and other pacific Islander Population

P0010035

White; Asian; Some other race population

P0010036

White; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010037

Black or African American; American Indian and Alaska Native; Asian Population

P0010038

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander population

P0010039

Black or African American; American Indian and Alaska Native; Some Other Race population

P0010040

Black or African American; Asian; Native Hawaiian and Other Pacific Islander Population

P0010041

Black or African American; Asian; Some Other Race population

P0010042

Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010043

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander Population

P0010044

American Indian and Alaska Native; Asian; Some Other Race population

P0010045

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010046

Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010046

Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population

P0010047

Population of four races

P0010048

White; Black or African American; American Indian and Alaska Native; Asian

P0010049

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0010050

White; Black or African American; American Indian and Alaska Native; Some Other Race

P0010051

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0010052

White; Black or African American; Asian; Some Other Race

P0010053

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010054

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0010055

White; American Indian and Alaska Native; Asian; Some Other Race

P0010056

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010057

White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010058

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0010059

Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0010060

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010061

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010062

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010063

Population of five races

P0010064

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0010065

White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0010066

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010067

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010068

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010069

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0010070

Population of six races

P0010071

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020001

Total:

P0020002

Hispanic or Latino

P0020003

Not Hispanic or Latino

P0020004

Population of one race

P0020005

White alone

P0020006

Black or african american alone

P0020007

American Indian or Alaskan Native alone

P0020008

Asian Alone

P0020009

Native Hawaiian and Other Pacific Islander alone

P0020010

Some other race alone

P0020011

Population of two or more races

P0020012

Population of two races

P0020013

White; Black or African American

P0020014

White; American Indian and Alaska Native

P0020015

White; Asian

P0020016

White; Native Hawaiian and Other Pacific Islander

P0020017

White; Some other race

P0020018

Black or African American; American Indian and Alaska Native

P0020019

Black or African American; Asian

P0020020

Black or African American; Native Hawaiian and Other Pacific Islander

P0020021

Black or African American; Some Other Race

P0020022

American Indian and Alaska Native; Asian

P0020023

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020024

American Indian and Alaska Native; Some Other Race

P0020025

Asian; Native Hawaiian and Other Pacific Islander

P0020026

Asian; Some Other Race

P0020027

Native Hawaiian and Other Pacific Islander; Some Other Race

P0020028

Population of three races

P0020029

White; Black or African American; American Indian and Alaska Native

P0020030

White; Black or African American; Asian

P0020031

White; Black or African American; Native Hawaiian and Other Pacific Islander

P0020032

White; Black or African American; Some Other Race

P0020033

White; American Indian and Alaska Native; Asian

P0020034

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020035

White; American Indian and Alaska Native; Some Other Race

P0020036

White; Asian; Native Hawaiian and Other Pacific Islander

P0020037

White; Asian; Some Other Race

P0020038

White; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020039

Black or African American; American Indian and Alaska Native; Asian

P0020040

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020021 Black or African American; Some Other Race
P0020041

Black or African American; American Indian and Alaska Native; Some Other Race

P0020042

Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0020043

Black or African American; Asian; Some Other Race

P0020044

Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020045

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020046

American Indian and Alaska Native; Asian; Some Other Race

P0020047

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020048

Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020049

Population of four races;

P0020050

White; Black or African American; American Indian and Alaska Native; Asian

P0020051

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0020052

White; Black or African American; American Indian and Alaska Native; Some Other Race

P0020053

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0020054

White; Black or African American; Asian; Some Other Race

P0020055

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020056

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020057

White; American Indian and Alaska Native; Asian; Some Other Race

P0020058

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020059

White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020060

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020041 Black or African American; American Indian and Alaska Native; Some Other Race
P0020061

Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0020062

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020063

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020064

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020065

Population of five races

P0020066

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0020067

White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race

P0020068

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020069

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020070

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020071

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0020072

Population of six races

P0020073

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race

P0030001

Total

P0030002

Population of one race

P0030003

White alone

P0030004

Black or african american alone

P0030005

American Indian and Alaska native alone

P0030006

Asian Alone

P0030007

Native Hawaiian and pacific islander alone

P0030008

Some other race alone

P0030009

Population of two or more races

P0030010

Population of two races

P0030011

White; Black or African American

P0030012

White; American Indian and Alaska Native

P0030013

White; Asian

P0030014

White; Native Hawaiian and Other Pacific Islander

P0030015

White; Some other race

P0030016

Black or African American; American Indian and Alaska Native

P0030017

Black or African American; Asian

P0030018

Black or African American; Native Hawaiian and Other Pacific Islander

P0030019

Black or African American; Some other race

P0030020

American Indian and Alaska Native; Asian

P0030021

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030022

American Indian and Alaska Native; Some other race

P0030023

Asian; Native Hawaiian and Other Pacific Islander

P0030024

Asian; Some other race

P0030025

Native Hawaiian and Other Pacific Islander; Some other race

P0030026

Population of three races

P0030027

White; Black or African American; American Indian and Alaska Native

P0030028

White; Black or African American; Asian

P0030029

White; Black or African American; Native Hawaiian and Other Pacific Islander

P0030030

White; Black or African American; Some other race

P0030031

White; American Indian and Alaska Native; Asian

P0030032

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030033

White; American Indian and Alaska Native; Some other race

P0030034

White; Asian; Native Hawaiian and Other Pacific Islander

P0030035

White; Asian; Some other race

P0030036

White; Native Hawaiian and Other Pacific Islander; Some other race

P0030037

Black or African American; American Indian and Alaska Native; Asian

P0030038

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030039

Black or African American; American Indian and Alaska Native; Some other race

P0030040

Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0030041

Black or African American; Asian; Some other race

P0030042

Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0030043

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030044

American Indian and Alaska Native; Asian; Some other race

P0030045

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030046

Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030047

Population of four races

P0030048

White; Black or African American; American Indian and Alaska Native; Asian

P0030049

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0030050

White; Black or African American; American Indian and Alaska Native; Some other race

P0030051

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0030052

White; Black or African American; Asian; Some other race

P0030053

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0030054

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030055

White; American Indian and Alaska Native; Asian; Some other race

P0030056

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030057

White; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030058

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030059

Black or African American; American Indian and Alaska Native; Asian; Some other race

P0030060

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030061

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030062

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030063

Population of five races

P0030064

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0030065

White; Black or African American; American Indian and Alaska Native; Asian; Some other race

P0030066

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0030067

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030068

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030069

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0030070

Population of six races

P0030071

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040001

Total

P0040002

Hispanic or latino

P0040003

Not hispanic or latino

P0040004

Population of one race

P0040005

White alone

P0040006

Black or african american alone

P0040007

American Indian and alaska native alone

P0040008

Asian Alone

P0040009

Native Hawaiian and pacific islander alone

P0040010

Some other race alone

P0040011

Population of two or more races

P0040012

Population of two races

P0040013

White; Black or African American

P0040014

White; American Indian and Alaska Native

P0040015

White; Asian

P0040016

White; Native Hawaiian and Other Pacific Islander

P0040017

White; Some other race

P0040018

Black or African American; American Indian and Alaska Native

P0040019

Black or African American; Asian

P0040020

Black or African American; Native Hawaiian and Other Pacific Islander

P0040021

Black or African American; Some other race

P0040022

American Indian and Alaska Native; Asian

P0040023

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040024

American Indian and Alaska Native; Some other race

P0040025

Asian; Native Hawaiian and Other Pacific Islander

P0040026

Asian; SOme other race

P0040027

Native Hawaiian and Other Pacific Islander; Some other race

P0040028

Population of three races

P0040029

White; Black or African American; American Indian and Alaska Native

P0040030

White; Black or African American; Asian

P0040031

White; Black or African American; Native Hawaiian and Other Pacific Islander

P0040032

White; Black or African American; Some other race

P0040033

White; American Indian and Alaska Native; Asian

P0040034

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040035

White; American Indian and Alaska Native; Some other race

P0040036

White; Asian; Native Hawaiian and Other Pacific Islander

P0040037

White; Asian; Some other race

P0040038

White; Native Hawaiian and Other Pacific Islander; Some other race

P0040039

Black or African American; American Indian and Alaska Native; Asian

P0040040

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040041

Black or African American; American Indian and Alaska Native; Some other race

P0040042

Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0040043

Black or african american; Asian; some other race

P0040044

Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0040045

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040046

American Indian and Alaska Native; Asian; Some other race

P0040047

American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040048

Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040049

Population of four races

P0040050

White; Black or African American; American Indian and Alaska Native; Asian

P0040051

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander

P0040052

White; Black or African American; American Indian and Alaska Native; Some other race

P0040053

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander

P0040054

White; Black or African American; Asian; Some other race

P0040055

White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race

P0040056

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040057

White; American Indian and Alaska Native; Asian; Some other race

P0040058

White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040059

White; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040060

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040061

Black or African American; American Indian and Alaska Native; Asian; Some other race

P0040062

Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040063

Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040064

American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040065

Population of five races

P0040066

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander

P0040067

White; Black or African American; American Indian and Alaska Native; Asian; Some other race

P0040068

White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race

P0040069

White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040070

White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040071

Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

P0040072

Popukation of six races

P0040073

White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race

H0010001

Total

H0010002

Occupied

H0010003

Vacant

P0050001

Total

P0050002

Instituonalized population

P0050003

Correctional facilities for adults

P0050004

Juvenile facilities

P0050005

Nursing facilities/Skilled-nursing facilities

P0050006

Other institutional facilities

P0050007

Noninstitutionalized population:

P0050008

College/University student housing

P0050009

Military quarters

P0050010

Other noninstitutional facilities

MTFCC20

MAT/TIGER Feature Class Code

geometry

Geometric data for the shape. (Point, Multipoint, Polyline, Polygon,etc)

alabamacounty20

alaskacounty20

arizonacounty20

arkansascounty20

californiacounty20

coloradocounty20

connecticutcounty20

delawarecounty20

dccounty20

floridacounty20

georgiacounty20

hawaiicounty20

idahocounty20

illinoiscounty20

indianacounty20

iowacounty20

kansascounty20

kentuckycounty20

louisianacounty20

mainecounty20

marylandcounty20

massachussetscounty20

michigancounty20

minnesotacounty20

mississippicounty20

missouricounty20

montanacounty20

nebraskacounty20

nevadacounty20

newhampshirecounty20

newjerseycounty20

newmexicocounty20

newyorkcounty20

northcarolinacounty20

northdakotacounty20

ohiocounty20

oklahomacounty20

oregoncounty20

pennsylvaniacounty20

puertoricocounty20

rhodeislandcounty20

southcarolinacounty20

southdakotacounty20

tennessecounty20

texascounty20

utahcounty20

vermontcounty20

virginiacounty20

washingtoncounty20

westvirginiacounty20

wisconsincounty20

wyomingcounty20

Format

An object of class sf (inherits from data.frame) with 67 rows and 400 columns.

An object of class sf (inherits from data.frame) with 30 rows and 400 columns.

An object of class sf (inherits from data.frame) with 15 rows and 400 columns.

An object of class sf (inherits from data.frame) with 75 rows and 400 columns.

An object of class sf (inherits from data.frame) with 58 rows and 400 columns.

An object of class sf (inherits from data.frame) with 64 rows and 400 columns.

An object of class sf (inherits from data.frame) with 8 rows and 400 columns.

An object of class sf (inherits from data.frame) with 3 rows and 400 columns.

An object of class sf (inherits from data.frame) with 1 rows and 400 columns.

An object of class sf (inherits from data.frame) with 67 rows and 400 columns.

An object of class sf (inherits from data.frame) with 159 rows and 400 columns.

An object of class sf (inherits from data.frame) with 5 rows and 400 columns.

An object of class sf (inherits from data.frame) with 44 rows and 400 columns.

An object of class sf (inherits from data.frame) with 102 rows and 400 columns.

An object of class sf (inherits from data.frame) with 92 rows and 400 columns.

An object of class sf (inherits from data.frame) with 99 rows and 400 columns.

An object of class sf (inherits from data.frame) with 105 rows and 400 columns.

An object of class sf (inherits from data.frame) with 120 rows and 400 columns.

An object of class sf (inherits from data.frame) with 64 rows and 400 columns.

An object of class sf (inherits from data.frame) with 16 rows and 400 columns.

An object of class sf (inherits from data.frame) with 24 rows and 400 columns.

An object of class data.frame with 14 rows and 398 columns.

An object of class sf (inherits from data.frame) with 83 rows and 400 columns.

An object of class sf (inherits from data.frame) with 87 rows and 400 columns.

An object of class sf (inherits from data.frame) with 82 rows and 400 columns.

An object of class sf (inherits from data.frame) with 115 rows and 400 columns.

An object of class sf (inherits from data.frame) with 56 rows and 400 columns.

An object of class sf (inherits from data.frame) with 93 rows and 400 columns.

An object of class sf (inherits from data.frame) with 17 rows and 400 columns.

An object of class sf (inherits from data.frame) with 10 rows and 400 columns.

An object of class sf (inherits from data.frame) with 21 rows and 400 columns.

An object of class sf (inherits from data.frame) with 33 rows and 400 columns.

An object of class sf (inherits from data.frame) with 62 rows and 400 columns.

An object of class sf (inherits from data.frame) with 100 rows and 400 columns.

An object of class sf (inherits from data.frame) with 53 rows and 400 columns.

An object of class sf (inherits from data.frame) with 88 rows and 400 columns.

An object of class sf (inherits from data.frame) with 77 rows and 400 columns.

An object of class sf (inherits from data.frame) with 36 rows and 400 columns.

An object of class sf (inherits from data.frame) with 67 rows and 400 columns.

An object of class sf (inherits from data.frame) with 78 rows and 400 columns.

An object of class sf (inherits from data.frame) with 5 rows and 400 columns.

An object of class sf (inherits from data.frame) with 46 rows and 400 columns.

An object of class sf (inherits from data.frame) with 66 rows and 400 columns.

An object of class data.frame with 95 rows and 398 columns.

An object of class data.frame with 254 rows and 398 columns.

An object of class sf (inherits from data.frame) with 29 rows and 400 columns.

An object of class sf (inherits from data.frame) with 14 rows and 400 columns.

An object of class sf (inherits from data.frame) with 133 rows and 400 columns.

An object of class sf (inherits from data.frame) with 39 rows and 400 columns.

An object of class sf (inherits from data.frame) with 55 rows and 400 columns.

An object of class sf (inherits from data.frame) with 72 rows and 400 columns.

An object of class data.frame with 23 rows and 398 columns.

Source

https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/